+++ pkg/pub_integration/lib/src/fake_pub_server_process.dart
@@ -66,7 +66,7 @@ class FakePubServerProcess {
       workingDirectory: pkgDir,
       environment: {
         // Because we read the consent email from stdout.
-        'DEBUG': 'fake_server pub.email',
+        'DEBUG': Platform.environment['DEBUG'] ?? 'fake_server pub.email',
       },
     );
     final instance = FakePubServerProcess._(port, process, coverageConfig);